# UTF-8 to Latin 1 (ARM)