--- a/benchmarks/src/apple_arm_events.h
+++ b/benchmarks/src/apple_arm_events.h
@@ -993,7 +993,6 @@ struct AppleEvents {
     // check permission
     int force_ctrs = 0;
     if (kpc_force_all_ctrs_get(&force_ctrs)) {
-      printf("Permission denied, xnu/kpc requires root privileges.\n");
       return (worked = false);
     }
     int ret;