On Apple M2 (LLVM 14) with the `french.utflatin8.txt` file.

```
$ sudo ./benchmarks/benchmark -P convert_utf8_to_latin1 -F ../unicode_lipsum/wikipedia_mars/french.utflatin8.txt
Password:
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 73.2
Using iconv version 267
testcases: 1
input detected as UTF8
current system detected as arm64
===========================
convert_utf8_to_latin1+arm64, input size: 440052, iterations: 30000, dataset: ../unicode_lipsum/wikipedia_mars/french.utflatin8.txt
   1.864 ins/byte,    1.005 cycle/byte,    3.420 GB/s (9.8 %),     3.439 GHz,    1.854 ins/cycle 
   1.897 ins/char,    1.023 cycle/char,    3.360 Gc/s (9.8 %)     1.02 byte/char 
convert_utf8_to_latin1+iconv, input size: 440052, iterations: 30000, dataset: ../unicode_lipsum/wikipedia_mars/french.utflatin8.txt
  45.452 ins/byte,    7.595 cycle/byte,    0.449 GB/s (2.4 %),     3.412 GHz,    5.984 ins/cycle 
  46.267 ins/char,    7.731 cycle/char,    0.441 Gc/s (2.4 %)     1.02 byte/char 
convert_utf8_to_latin1+icu, input size: 440052, iterations: 30000, dataset: ../unicode_lipsum/wikipedia_mars/french.utflatin8.txt
   9.061 ins/byte,    2.384 cycle/byte,    1.435 GB/s (1.8 %),     3.421 GHz,    3.801 ins/cycle 
   9.223 ins/char,    2.426 cycle/char,    1.410 Gc/s (1.8 %)     1.02 byte/char 
convert_utf8_to_latin1_with_errors+arm64, input size: 440052, iterations: 30000, dataset: ../unicode_lipsum/wikipedia_mars/french.utflatin8.txt
   1.897 ins/byte,    1.039 cycle/byte,    3.308 GB/s (2.8 %),     3.438 GHz,    1.825 ins/cycle 
   1.931 ins/char,    1.058 cycle/char,    3.249 Gc/s (2.8 %)     1.02 byte/char 
```

With pure ASCII:

```
$ sudo ./benchmarks/benchmark -P convert_utf8_to_latin1 -F ../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 73.2
Using iconv version 267
testcases: 1
input detected as UTF8
current system detected as arm64
===========================
convert_utf8_to_latin1+arm64, input size: 86940, iterations: 30000, dataset: ../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   0.436 ins/byte,    0.081 cycle/byte,   83.516 GB/s (11.6 %),     6.755 GHz,    5.389 ins/cycle 
   0.436 ins/char,    0.081 cycle/char,   83.516 Gc/s (11.6 %)     1.00 byte/char 
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf8_to_latin1+iconv, input size: 86940, iterations: 30000, dataset: ../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  46.167 ins/byte,    7.414 cycle/byte,    0.462 GB/s (0.6 %),     3.429 GHz,    6.227 ins/cycle 
  46.167 ins/char,    7.414 cycle/char,    0.462 Gc/s (0.6 %)     1.00 byte/char 
convert_utf8_to_latin1+icu, input size: 86940, iterations: 30000, dataset: ../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   9.221 ins/byte,    2.050 cycle/byte,    1.698 GB/s (0.6 %),     3.481 GHz,    4.497 ins/cycle 
   9.221 ins/char,    2.050 cycle/char,    1.698 Gc/s (0.6 %)     1.00 byte/char 
convert_utf8_to_latin1_with_errors+arm64, input size: 86940, iterations: 30000, dataset: ../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   0.436 ins/byte,    0.082 cycle/byte,   80.277 GB/s (7.5 %),     6.556 GHz,    5.342 ins/cycle 
   0.436 ins/char,    0.082 cycle/char,   80.277 Gc/s (7.5 %)     1.00 byte/char 
```

Fixes https://github.com/simdutf/simdutf/issues/277