@lemire 2023-09-06
We probably can do better performance-wise... I might revisit it, but we are already 'fast'.